path following implemented, with testing controls
