+++ TraceViewsViewController.swift
@@ -8,6 +8,7 @@
 
 import AppKit
 import RxSwift
+import QuartzCore
 
 func map<S, T>(f: @escaping (S) -> T) -> ([S]) -> [T] {
     return { $0.map(f) }
@@ -434,158 +435,6 @@ class TraceView: NSView {
     private func ballPath(at point: NSPoint, radius r: Scalar) -> NSBezierPath {
         return NSBezierPath(ovalIn: NSRect(origin: NSPoint(x: -r, y: -r) + point, size: NSSize(width: 2*r, height: 2*r)))
     }
-}
-
-let n = 60
-let c = NSPoint(x: 700, y: 400)
-
-let speed: Scalar = 5
-
-class PursuitView: NSView {
-    var position = NSPoint.zero
-    var yaw: Scalar = π/4
-
-    let basePathPoints = (0..<n).map { Scalar($0)*2*π/Scalar(n) }.map { NSPoint(x: sin($0), y: cos($0)) }
-    var pathRadius: Scalar = 300
-
-    var pathPoints: [NSPoint] { return basePathPoints.map { c + pathRadius*$0} }
-
-    var index = 0
-    var target: NSPoint { return pathPoints[index] }
-
-    var arcCenter = NSPoint.zero
-    var arcRadius: Scalar = 1000000
-    var yawRate: Scalar = 0
-
-    var searchRadius: Scalar = 200
-
-    var trail = [NSPoint]()
-
-    required init?(coder: NSCoder) {
-        super.init(coder: coder)
-
-        acceptsTouchEvents = true
-    }
-
-    override func mouseDown(with event: NSEvent) {
-        processMouseEvent(event)
-    }
-
-    override func mouseDragged(with event: NSEvent) {
-        processMouseEvent(event)
-    }
-
-    private func processMouseEvent(_ event: NSEvent) {
-        position = convert(event.locationInWindow, from: nil)
-        next()
-        trail = []
-        setNeedsDisplay(bounds)
-    }
-
-    func radiusChanged(r: Scalar) {
-        pathRadius = r
-    }
-
-    func next() {
-        updateTarget()
-        updateArc()
-        updateYawRate()
-        updatePosition()
-        trail.append(position)
-    }
-
-    func updateTarget() {
-        var i = index
-
-        while (pathPoints[i] - position).norm < searchRadius {
-            i = (i + 1) % n
-        }
-        index = i
-    }
-
-    func updateArc() {
-        let attitude = NSPoint(phi: yaw, r: 1)
-        let phi = attitude.signedAngle(to: target - position)
-        arcRadius = (target - position).norm/sqrt(2*(1 - cos(2*phi)))
-        arcCenter = position + arcRadius*attitude.rotated(by: (phi.sign == .plus ? 1 : -1)*π/2)
-    }
-
-    func updateYawRate() {
-
-    }
-
-    func updatePosition() {
-
-    }
-
-    override func draw(_ dirtyRect: NSRect) {
-        // Draw search area
-        NSColor(calibratedRed: 0.8, green: 0.8, blue: 0.8, alpha: 1).set()
-        ballPath(at: position, radius: searchRadius).fill()
-
-        // Draw kite path
-        NSColor.lightGray.set()
-        ballPath(at: c, radius: pathRadius).stroke()
-
-        // Draw target
-        NSColor.blue.set()
-        ballPath(at: target, radius: 5).stroke()
-
-        // Draw trail
-        NSColor.gray.set()
-        makePath(points: trail)?.stroke()
-
-        // Draw position
-        NSColor.orange.set()
-        ballPath(at: position, radius: 5).fill()
-        makePath(lines: [(position, position + NSPoint(phi: yaw, r: 50))]).stroke()
-
-        // Draw arc
-        NSColor.purple.set()
-        let arc = NSBezierPath()
-        let startAngle = 180*(position - arcCenter).phi/π
-        let endAngle = 180*(target - arcCenter).phi/π
-
-        let clockWise = (target - position).signedAngle(to: NSPoint(phi: yaw, r: 1)) > 0
-        arc.appendArc(withCenter: arcCenter, radius: arcRadius, startAngle: startAngle, endAngle: endAngle, clockwise: clockWise)
-        arc.lineWidth = 2
-        arc.stroke()
-
-        ballPath(at: arcCenter, radius: 5).fill()
-
-        //        makePath(lines: [(position, position + 10*)]) // draw yawrate
-    }
-
-    private func ballPath(at point: NSPoint, radius r: Scalar) -> NSBezierPath {
-        return NSBezierPath(ovalIn: NSRect(origin: NSPoint(x: -r, y: -r) + point, size: NSSize(width: 2*r, height: 2*r)))
-    }
-
-    private func makePath(lines: [(start: NSPoint, end: NSPoint)]) -> NSBezierPath {
-        let p = NSBezierPath()
-        p.lineWidth = 2
-
-        for line in lines {
-            p.move(to: line.start)
-            p.line(to: line.end)
-        }
-        return p
-    }
-
-    private func makePath(points: [NSPoint]) -> NSBezierPath? {
-        guard let first = points.first else { return nil }
-
-        let p = NSBezierPath()
-        p.lineWidth = 2
-
-        p.move(to: first)
-
-        for point in points.dropFirst() {
-            p.line(to: point)
-        }
-
-        return p
-    }
-
 }
 
 class SphereView: NSView {
